code/backend/src/api/requirements.txt
psycopg2-binary
uvicorn
fastapi
asyncpg
websockets
httpx